chore: relicense to Apache-2.0
This commit relicenses all of the codebase to Apache-2.0. The logic
behind this is that so much of the codebase is Apache-2.0, from
before the fork, that it doesn't make sense to dual-license the
repository. All projects under this org going forward will be GPL,
but stencil itself will stay as Apache-2.0

For proof of authors who contributed during this time's agreement
to relicense, see the PR associated with this change.