feat: support library templates (#64)
This commit adds support for special "library" templates that can never
generate files. The intent for these is to provide helper
functions/sub-templates for other templates to call. A libary template
is any template file that has the `.library.tpl` extension.

Adds tests ensuring that this functionality works as expected as well as
updates the documentation to reflect the introduction of them.