feat: Adding Directory Manifests, currently only for templated dir renaming (#63)
Java/Kotlin projects tend to need to name entire folders based on a
package name, so to reduce a ton of error-prone boilerplate stencil file
management in these sorts of template modules, we are introducing the
notion of a directory manifest. For now, it only contains a template
that will be rendered and used as the new name for the directory for all
templates rendered underneath it.