fix!(create module): only create stencil.yaml for now
Fixes #46 by way of removing most of the functionality in `create
module` for now, due to it being unclear what we'll carry over from the
existing template-base code. A future commit will improve this.

Adds a small test framework for testing commands which this adds a test
for `create module`.