feat: add slogext package and switch to use it
This commit adds a `slogext` package with the purpose of providing a
generic logging package, powered by `slog`[^1], with `charm/log`[^2] as
the underlying handler. This enables us to have an easier time switching
logging libraries out if we ever need to do so again.

I also added the notion of a test logger that will enable us to do test
specific things if we so desire (e.g., pipe to `t.Log`, automatic
`test.name` setting, etc.). I switched all tests to use this.

Misc changes made:
- Added a Apache-2.0 license header to the snippets replacing the stable
  AGPL one there previously.
- Fixed a bug in the stenciltest error handling that would cause
  expecting an error to never actually work.

[^1]: https://pkg.go.dev/log/slog
[^2]: https://pkg.go.dev/github.com/charmbracelet/log