refactor: rename extensions to nativeext
Renames `extensions` to `nativeext` and starts cleaning up the module
resolving logic to be less all over the place and to better integrate
the native extensions code with the rest of the module resolving logic.

This also switches from `go-git` to running `git` commands directly to
ease the system authentication story.