refactor(modules): add module resolver (#82)
Adds a module resolver library for the purposes of replacing the current
version resolver used by the `modules` package. This does not hook it up
yet, but provides us a scalable foundation for resolving versions.

Future PR(s) will change `stencil` to default to using versions found in
the lockfile, as well as adding a `stencil upgrade` command to upgrade
them.

The underlying resolver introduced in this PR supports simple constraint
functionality to allow flexibility in its usage in `stencil`, but there
is currently no plans to expose this functionality. Tests have been
added to ensure that the following behaviour is followed:

* If a criteria contains a branch, that branch should always be used.
  * **Reason**: Branches cannot be compared and usually are being used
    for testing purposes, so we allow those to always win. Future code
    will restrict them to the top-level module only to prevent
    sub-dependencies from being released with branch dependencies.
* If a criteria contains a prerelease string, only one distinct value
  should be used.
  * **Reason**: We cannot compare `v1.0.0-alpha.1` against
    `v1.0.0-beta.1` generally speaking, so we instead only allow
    pre-release criterias to opt into a specific pre-release track.
    Released versions till rank higher.
* Only one branch can be used per-criteria, since we cannot depend on
  multiple branches for the same module.