feat!: use new resolver, remove stenciltest, --frozen-lockfile
Uses the new versions resolver under the hood for the `version` field in
the modules list.

Supports constraints and branches.

Future PR will add 'back' frozen lockfile in the sense that it'll be the
default, and `stencil upgrade` will upgrade versions. If a module
explictly requests a version it can do so by not opting to use a
constraint, but by default a version == pin inside of a manifest.

Major versions prompting has also been removed.