refactor(box)!: remove box configuration (#106)
The current box configuration implementation requires `gobox`, which
we're almost done removing from the dependency chain. As such, I'd like
to remove it.

My reasoning for doing so is that the current experience of `box`
creates gaps with open source repositories that are owned by an org,
causing those users to be unable to run `stencil` in those repositories,
normally resulting in an inability to build/run them.

I'd like to add box back later in another form, but for now it's best to
remove it and add it back later.

As far as I know, the current implementation also doesn't work in this
fork.