fix: properly format files and re-run formatters on the repo
Runs formatters on the repository as they were meant to be configured as
well adds `linters` Github Action group that ensures that formatters
were ran as well as code/doc generators.

Enables tabs instead of spaces across the entire repository because they
use less space and any modern editor should be configured to display
them to their liking. The only exception is YAML which has to be spaces.
We output `2` for the least amount as well as it being my preference
(again, configure your editor if you haven't already, if you don't like
2).