fix: reresolve modules/replacements when modified in manifest
Previously, if anything was changed in `stencil.yaml` `stencil upgrade`
needed to be ran for the behaviour to have any effect. This is undesired
because `stencil upgrade` upgrades everything and isn't the desired
behaviour when changing this file.

Instead, we now determine when either the module version or a
replacement is added and force a resolve of just that module, without
considering it to be a lockfile violation, because the user changed that
file.

Resolves #114