feat: use jaredallard/vcs to support Gitlab and almost move off gobox
Changes us to use https://github.com/jaredallard/vcs instead of a mix of
our own copies and gobox for release fetching. This leaves us with just
archive extraction from gobox, which will be removed soon.

This is mostly a minor import URL change since nothing was changed
between the various versions in that repo.