feat(create-module): use stencil-golang to create modules
Adds support for generating modules with a `manifest.yaml` as well as
populating `service.yaml` using `rgst-io/stencil-golang` in library
mode, unless specified to create a native extension.

stencil-golang in library mode has full CI, and is used mainly only for
releasing. Eventually, the release logic will be moved out and turned
into some framework (or use some other one) since it is a bit awkward to
depend on that for non-Go repositories.