feat(codegen): support calling functions across modules
Functions are a new concept that are only available to library
templates. A function is a template, declared through `define` that can
be called across modules.

Functions must return a value through the `return` function, which is
only callable inside of a function template. A function template must be
exported through the `module.Export`. Once exported, it can be called
using `module.Call "module-name.FunctionName"`.

Functions behave the same way as templates called through
`ApplyTemplate` do, where data is passed on `.Data` of the `Values`
object.