feat(codegen): run n passes
Moves to a fixed two pass system to an `n` pass system, which defaults
to 20 max.

This is done by moving state into a single struct, which we hash between
runs. Onces state has not changed, we consider the first pass to be
done. This is a pretty naive approach, in the future we could possibly
filter down the list of templates to re-render.