docs: rename generate function docs to fix URLs (#154)
For some reason vitepress' router seems to break if pages have multiple
dots in the name of the file. So, we replace it with `_` instead and
adjust the title to make it work properly.

Seems to work locally but won't know for sure until I get a cloudflare
page running.