fix: Cross-module calls should execute in the context of the callee (#177)
The original design was to use the context of the caller in the
functions being called, but after we tried to implement a function call
this way realized that it was inverting ownership from usual function
ownership. The idea was to be able to use helper functions in the
context of the caller, but the reality is that you want to pass in any
needed context to the exported function, just like in standard
programming. This way the function owner can have "private"
state/context inside the module that the function exposes a part of, per
standard modules in general software engineering.