fix(modules): better error when failing to resolve module(s)
Changes the resolution error[^1] to be more detailed when a repository
cannot be accessed (e.g., invalid token) vs a resolution (criteria)
error.

Previously, we'd always return resolution history when we failed, but
the error could be related to an invalid token instead of the history.
There would be no way to know this happened. Now, we conditionally
display the history (based on classes of errors), but always return the
resolver error. Updating the `vcs` library now also provides more
information in the returned error that we're displaying.

[^1]: ```
INFO stencil 0.15.0
INFO Fetching dependencies
ERRO failed to run
 error=
 │ failed to resolve module 'github.com/rgst-io/not-real-repo' with constraints
 │ └─ a-stencil-module (top-level) wants >=0.0.0
```