fix(modules): display more information on repo not found
Adds more details when the error is due to remotes being unable to be
found. This could be a credentials issue, so it's worth highlighting it
in the error message.

As part of this, I was reminded that we use `git` only for credentials
here, which should probably be switched to the token library in the
future. We don't do this right now since its hard to configure host git
on the fly to use the system-wide token, but it's worth considering to
unify token sources in the future.