fix(nativeext): race condition on parallel ext dl
Fixes a possible race condition with downloading native extensions when
running multiple stencil instances at the same time using the same
modules.

We now lock the cache directory and check for it at extension download.
Unfortunately, this blocks until we're able to obtain a lock so it's
possible this could hang as well as have bad UX when this occurs.