feat(template): Rename ApplyTemplate to Include (#220)
Renames `stencil.ApplyTemplate` to `Include`. `ApplyTemplate` has always
been a confusing name for a function that returns a templates execution
output. So, we rename it to `Include`. `ApplyTemplate` is kept around,
but deprecated. A warning is returned when a module calls it.

Due to `text/template` limitations, we cannot include the line that
called `ApplyTemplate` in the warning.