feat: Adding file.MigrateTo, a system for migrating old files to new templates
Do you want to move a file around in your target repositories, which already has customizations in blocks that you don't want to lose?  MigrateTo now moves this file for you, letting the new template pick it up.  The next pass of the template rendering will then pick up the migrated file in the new location under the new template and render any updates with the new template, as if the user had moved the file themselves.