feat: Adding file.BlockI feature for more-concise unused blocks
This uses the nuance that go-template blocks opened with {{- contract whitespace back to the end of the start of the block line, so if the block is empty you can return "" and have no extra empty newline between the start and end of the block.  If it has contents, the BlockI handler first prefixes the result with a newline so that it behaves like a standard block.