feat: add stencil module test framework (#321)
Adds a stencil testing framework under the new `stencil module test`
command.

The framework works by creating directories under the `tests` folder,
like so:

```bash
tests
--> SmokeTest
-----> stencil.yaml
```

`stencil` is then ran using the `stencil.yaml` in that folder, in a
temporary working
directory. If stencil renders with no error, a success state is
returned, otherwise a
failure state.

Optionally, there is now a new `testing` block added to `stencil.yaml`
that allows
controlling how tests are run against this manifest. For now, it only
includes
running a set of commands against the rendered state, as `validators`.
For
example, this could be something that checks a configuration is valid:

```yaml
testing:
  validators:
  - circleci validate
```

Tests have their own instance of a stencil generator, but are ran
in-process so it
is more efficient than running `stencil` yourself. Futhermore, the vcs
resolver
is cached for the lifetime of the stencil test run, caching any version
look-ups,
which tend to be the bulk of the time spent by stencil before rendering
templates.

Tests are currently ran serially, but eventually support will be added
to run them
in parallel. 

Related to #6